% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{add_link}
\alias{add_link}
\title{add link}
\usage{
add_link(url_path, url_text)
}
\arguments{
\item{url_path}{string, url}

\item{url_text}{string, url link name}
}
\value{
html url link
}
\description{
writes html for named url link to r markdown
}
\examples{
add_link("https://www.crumplab.com","crumplab.com")

}
