% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{add_link_icon}
\alias{add_link_icon}
\title{add link and icon}
\usage{
add_link_icon(url_path, url_text, icon_class)
}
\arguments{
\item{url_path}{string, the url}

\item{url_text}{string, name}

\item{icon_class}{string, font awesome icon class}
}
\value{
html for link with icon and text
}
\description{
writes html for named url link to r markdown with font awesome icon
}
\examples{
add_link_icon("https://www.crumplab.com","crumplab.com","fas fa-globe")

}
