% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{bib2annotate}
\alias{bib2annotate}
\title{Convert bib entries to annotation entries for R markdown}
\usage{
bib2annotate(cite_key, bib_file, csl_file = "apa.csl", header_level = "##")
}
\arguments{
\item{cite_key}{character vector, a vector of citation keys from a .bib file}

\item{bib_file}{string, path to .bib file}

\item{csl_file}{string, name of csl style file, default is apa.csl}

\item{header_level}{string of hashtags, number of hashtags to set header level, default is "##"}
}
\value{
markdown The markdown to be written into the R Markdown document. Use with results="asis"
}
\description{
This is a convenience wrapper function for printing full csl-style bibliographic entries in the middle of an R Markdown document.
}
