% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{bib2md}
\alias{bib2md}
\title{Convert bib entries to markdown for printing citations in text}
\usage{
bib2md(cite_key, bib_file, csl_file = "apa.csl")
}
\arguments{
\item{cite_key}{character vector, a vector of citation keys from a .bib file}

\item{bib_file}{string, path to .bib file}

\item{csl_file}{sting, name of csl style file, default is apa.csl}
}
\value{
markdown The markdown to be written into the R Markdown document. Suggested use case is to call this function in an R code chunk with results="asis"
}
\description{
This is a convenience wrapper function for printing full csl-style bibliographic entries in the middle of an R Markdown document.
}
