% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{bib_2_pub_list}
\alias{bib_2_pub_list}
\title{Generate a publication list from a bib file}
\usage{
bib_2_pub_list(bib, yml, pdf_dir, base_url_to_pdfs)
}
\arguments{
\item{bib}{a .bib file}

\item{yml}{a formatted .yml file see example}

\item{pdf_dir}{relative path to directory containing pdfs}

\item{base_url_to_pdfs}{url to path containing pdfs}
}
\value{
publication list with links
}
\description{
This function takes a .bib file as input, along with a specially formatted .yml file and embeds a publication list into an R Markdown document.
}
\examples{
\dontrun{
bib_2_pub_list("Crump.bib",
"Crump.yml",
"pkgdown/assets/Crump/files",
"https://www.crumplab.com/Crump/files")
}
}
