% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{crumplab_blog}
\alias{crumplab_blog}
\title{build blog pages}
\usage{
crumplab_blog()
}
\value{
builds all vignettes and tweaks blog index
}
\description{
Builds all vignettes using\code{pkgdown::build_articles()} and then tweaks the list of articles on the blog page by running \code{crumplab_blog_index()}.
}
\examples{
\dontrun{
crumplab_blog()
}
}
