% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{crumplab_blog_index}
\alias{crumplab_blog_index}
\title{create index.html for articles blog page}
\usage{
crumplab_blog_index(pkg = ".", override = yaml::read_yaml("_blog.yml"))
}
\arguments{
\item{override}{}
}
\value{

}
\description{
create index.html for articles blog page
}
