% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{crumplab_site}
\alias{crumplab_site}
\title{build site and then tweak}
\usage{
crumplab_site()
}
\value{
The pkgdown website
}
\description{
Equivalent of \code{pkgdown::build_site()} but tweaks the blog index page at the end by running \code{crumplab_blog_index()}
}
\examples{
\dontrun{
crumplab_site()
}
}
