% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{non_recursive_moving}
\alias{non_recursive_moving}
\title{Outlier exclusion using non-recursive procedure}
\usage{
non_recursive_moving(rts)
}
\arguments{
\item{rts}{a vector of reaction times}
}
\value{
list containing original rts, restricted rts with outlier removed, and proportion removed
}
\description{
Reproduces the non-recursive outlier exclusion procedure for reaction times from Van Selst, M., & Jolicoeur, P. (1994). A solution to the effect of sample size on outlier elimination. The Quarterly Journal of Experimental Psychology Section A, 47(3), 631-650.
}
\examples{
modified_recursive_moving(c(500,501,502,503,600))
}
