% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{vector numeric}
}
\value{
numeric, standard error of the mean
}
\description{
Standard error of the mean
}
\examples{
sem(c(1,2,3,4,5))
}
